data/requirements.txt
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
wordcloud>=1.9.0
scikit-learn>=1.1.0
python-dotenv>=0.19.0
nltk>=3.7
plotly>=5.0.0 